Break deploy commands down into distinct files
