+++ src/server/pkg/deploy/cmds/cmds.go
@@ -4,28 +4,17 @@ import (
 	"bufio"
 	"bytes"
 	"encoding/base64"
-	"encoding/hex"
 	"fmt"
-	"io"
 	"io/ioutil"
-	"math/rand"
-	"net/http"
 	"net/url"
 	"os"
-	"path"
 	"regexp"
-	"sort"
 	"strconv"
 	"strings"
 	"time"
 
-	"github.com/pachyderm/pachyderm/src/client"
-	"github.com/pachyderm/pachyderm/src/client/auth"
-	"github.com/pachyderm/pachyderm/src/client/enterprise"
 	"github.com/pachyderm/pachyderm/src/client/pkg/config"
 	"github.com/pachyderm/pachyderm/src/client/pkg/errors"
-	"github.com/pachyderm/pachyderm/src/client/pkg/grpcutil"
-	"github.com/pachyderm/pachyderm/src/client/pkg/helm"
 	"github.com/pachyderm/pachyderm/src/client/version"
 	"github.com/pachyderm/pachyderm/src/server/pkg/cmdutil"
 	"github.com/pachyderm/pachyderm/src/server/pkg/deploy"
@@ -33,10 +22,7 @@ import (
 	"github.com/pachyderm/pachyderm/src/server/pkg/deploy/images"
 	_metrics "github.com/pachyderm/pachyderm/src/server/pkg/metrics"
 	"github.com/pachyderm/pachyderm/src/server/pkg/obj"
-	"github.com/pachyderm/pachyderm/src/server/pkg/serde"
-	clientcmd "k8s.io/client-go/tools/clientcmd/api/v1"
 
-	docker "github.com/fsouza/go-dockerclient"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
@@ -50,236 +36,7 @@ var (
 const (
 	defaultDashImage   = "pachyderm/dash"
 	defaultDashVersion = "0.5.48"
-
-	defaultIDEHubImage  = "pachyderm/ide-hub"
-	defaultIDEUserImage = "pachyderm/ide-user"
-
-	defaultIDEVersion      = "1.1.0"
-	defaultIDEChartVersion = "0.9.1" // see https://jupyterhub.github.io/helm-chart/
-
-	ideNotes = `
-Thanks for installing the Pachyderm IDE!
-
-It may take a few minutes for all of the pods to spin up. If you have kubectl
-access, you can check progress with:
-
-  kubectl get pod -l release=pachyderm-ide
-
-Once all of the pods are in the 'Ready' status, you can access the IDE in the
-following manners:
-
-* If you're on docker for mac, it should be accessible on 'localhost'.
-* If you're on minikube, run 'minikube service proxy-public --url' -- one or
-  both of the URLs printed should reach the IDE.
-* If you're on a cloud deployment, use the external IP of
-  'kubectl get service proxy-public'.
-
-For more information about the Pachyderm IDE, see these resources:
-
-* Our how-tos: https://docs.pachyderm.com/latest/how-tos/use-pachyderm-ide/
-* The Z2JH docs, which the IDE builds off of:
-  https://zero-to-jupyterhub.readthedocs.io/en/latest/
-`
 )
-
-func kubectl(stdin io.Reader, context *config.Context, args ...string) error {
-	var environ []string = nil
-	if context != nil {
-		tmpfile, err := ioutil.TempFile("", "transient-kube-config-*.yaml")
-		if err != nil {
-			return errors.Wrapf(err, "failed to create transient kube config")
-		}
-		defer os.Remove(tmpfile.Name())
-
-		config := clientcmd.Config{
-			Kind:           "Config",
-			APIVersion:     "v1",
-			CurrentContext: "pachyderm-active-context",
-			Contexts: []clientcmd.NamedContext{
-				clientcmd.NamedContext{
-					Name: "pachyderm-active-context",
-					Context: clientcmd.Context{
-						Cluster:   context.ClusterName,
-						AuthInfo:  context.AuthInfo,
-						Namespace: context.Namespace,
-					},
-				},
-			},
-		}
-
-		var buf bytes.Buffer
-		if err := encoder("yaml", &buf).Encode(config); err != nil {
-			return errors.Wrapf(err, "failed to encode config")
-		}
-
-		tmpfile.Write(buf.Bytes())
-		tmpfile.Close()
-
-		kubeconfig := os.Getenv("KUBECONFIG")
-		if kubeconfig == "" {
-			home, err := os.UserHomeDir()
-			if err != nil {
-				return errors.Wrapf(err, "failed to discover default kube config: could not get user home directory")
-			}
-			kubeconfig = path.Join(home, ".kube", "config")
-			if _, err = os.Stat(kubeconfig); os.IsNotExist(err) {
-				return errors.Wrapf(err, "failed to discover default kube config: %q does not exist", kubeconfig)
-			}
-		}
-		kubeconfig = fmt.Sprintf("%s%c%s", kubeconfig, os.PathListSeparator, tmpfile.Name())
-
-		// note that this will override `KUBECONFIG` (if it is already defined) in
-		// the environment; see examples under
-		// https://golang.org/pkg/os/exec/#Command
-		environ = os.Environ()
-		environ = append(environ, fmt.Sprintf("KUBECONFIG=%s", kubeconfig))
-
-		if stdin == nil {
-			stdin = os.Stdin
-		}
-	}
-
-	ioObj := cmdutil.IO{
-		Stdin:   stdin,
-		Stdout:  os.Stdout,
-		Stderr:  os.Stderr,
-		Environ: environ,
-	}
-
-	args = append([]string{"kubectl"}, args...)
-	return cmdutil.RunIO(ioObj, args...)
-}
-
-// Generates a random secure token, in hex
-func generateSecureToken(length int) string {
-	b := make([]byte, length)
-	if _, err := rand.Read(b); err != nil {
-		return ""
-	}
-	return hex.EncodeToString(b)
-}
-
-// Return the appropriate encoder for the given output format.
-func encoder(output string, w io.Writer) serde.Encoder {
-	if output == "" {
-		output = "json"
-	} else {
-		output = strings.ToLower(output)
-	}
-	e, err := serde.GetEncoder(output, w,
-		serde.WithIndent(2),
-		serde.WithOrigName(true),
-	)
-	if err != nil {
-		cmdutil.ErrorAndExit(err.Error())
-	}
-	return e
-}
-
-func kubectlCreate(dryRun bool, manifest []byte, opts *assets.AssetOpts) error {
-	if dryRun {
-		_, err := os.Stdout.Write(manifest)
-		return err
-	}
-	// we set --validate=false due to https://github.com/kubernetes/kubernetes/issues/53309
-	if err := kubectl(bytes.NewReader(manifest), nil, "apply", "-f", "-", "--validate=false", "--namespace", opts.Namespace); err != nil {
-		return err
-	}
-
-	fmt.Println("\nPachyderm is launching. Check its status with \"kubectl get all\"")
-	if opts.DashOnly || !opts.NoDash {
-		fmt.Println("Once launched, access the dashboard by running \"pachctl port-forward\"")
-	}
-	fmt.Println("")
-
-	return nil
-}
-
-// findEquivalentContext searches for a context in the existing config that
-// references the same cluster as the context passed in. If no such context
-// was found, default values are returned instead.
-func findEquivalentContext(cfg *config.Config, to *config.Context) (string, *config.Context) {
-	// first check the active context
-	activeContextName, activeContext, _ := cfg.ActiveContext(false)
-	if activeContextName != "" && to.EqualClusterReference(activeContext) {
-		return activeContextName, activeContext
-	}
-
-	// failing that, search all contexts (sorted by name to be deterministic)
-	contextNames := []string{}
-	for contextName := range cfg.V2.Contexts {
-		contextNames = append(contextNames, contextName)
-	}
-	sort.Strings(contextNames)
-	for _, contextName := range contextNames {
-		existingContext := cfg.V2.Contexts[contextName]
-
-		if to.EqualClusterReference(existingContext) {
-			return contextName, existingContext
-		}
-	}
-
-	return "", nil
-}
-
-func contextCreate(namePrefix, namespace, serverCert string) error {
-	kubeConfig, err := config.RawKubeConfig()
-	if err != nil {
-		return err
-	}
-	kubeContext := kubeConfig.Contexts[kubeConfig.CurrentContext]
-
-	clusterName := ""
-	authInfo := ""
-	if kubeContext != nil {
-		clusterName = kubeContext.Cluster
-		authInfo = kubeContext.AuthInfo
-	}
-
-	cfg, err := config.Read(false)
-	if err != nil {
-		return err
-	}
-
-	newContext := &config.Context{
-		Source:      config.ContextSource_IMPORTED,
-		ClusterName: clusterName,
-		AuthInfo:    authInfo,
-		Namespace:   namespace,
-		ServerCAs:   serverCert,
-	}
-
-	equivalentContextName, equivalentContext := findEquivalentContext(cfg, newContext)
-	if equivalentContext != nil {
-		cfg.V2.ActiveContext = equivalentContextName
-		equivalentContext.Source = newContext.Source
-		equivalentContext.ClusterDeploymentID = ""
-		equivalentContext.ServerCAs = newContext.ServerCAs
-		return cfg.Write()
-	}
-
-	// we couldn't find an existing context that is the same as the new one,
-	// so we'll have to create it
-	newContextName := namePrefix
-	if _, ok := cfg.V2.Contexts[newContextName]; ok {
-		newContextName = fmt.Sprintf("%s-%s", namePrefix, time.Now().Format("2006-01-02-15-04-05"))
-	}
-
-	cfg.V2.Contexts[newContextName] = newContext
-	cfg.V2.ActiveContext = newContextName
-	return cfg.Write()
-}
-
-// containsEmpty is a helper function used for validation (particularly for
-// validating that creds arguments aren't empty
-func containsEmpty(vals []string) bool {
-	for _, val := range vals {
-		if val == "" {
-			return true
-		}
-	}
-	return false
-}
 
 func standardDeployCmds() []*cobra.Command {
 	var commands []*cobra.Command
@@ -1011,158 +768,7 @@ If <object store backend> is \"s3\", then the arguments are:
 func Cmds() []*cobra.Command {
 	commands := standardDeployCmds()
 
-	var lbTLSHost string
-	var lbTLSEmail string
-	var dryRun bool
-	var outputFormat string
-	var jupyterhubChartVersion string
-	var hubImage string
-	var userImage string
-	deployIDE := &cobra.Command{
-		Short: "Deploy the Pachyderm IDE.",
-		Long:  "Deploy a JupyterHub-based IDE alongside the Pachyderm cluster.",
-		Run: cmdutil.RunFixedArgs(0, func(args []string) (retErr error) {
-			cfg, err := config.Read(false)
-			if err != nil {
-				return err
-			}
-			_, activeContext, err := cfg.ActiveContext(true)
-			if err != nil {
-				return err
-			}
-
-			c, err := client.NewOnUserMachine("user")
-			if err != nil {
-				return errors.Wrapf(err, "error constructing pachyderm client")
-			}
-			defer c.Close()
-
-			enterpriseResp, err := c.Enterprise.GetState(c.Ctx(), &enterprise.GetStateRequest{})
-			if err != nil {
-				return errors.Wrapf(grpcutil.ScrubGRPC(err), "could not get Enterprise status")
-			}
-
-			if enterpriseResp.State != enterprise.State_ACTIVE {
-				return errors.New("Pachyderm Enterprise must be enabled to use this feature")
-			}
-
-			authActive, err := c.IsAuthActive()
-			if err != nil {
-				return errors.Wrapf(grpcutil.ScrubGRPC(err), "could not check whether auth is active")
-			}
-			if !authActive {
-				return errors.New("Pachyderm auth must be enabled to use this feature")
-			}
-
-			whoamiResp, err := c.WhoAmI(c.Ctx(), &auth.WhoAmIRequest{})
-			if err != nil {
-				return errors.Wrapf(grpcutil.ScrubGRPC(err), "could not get the current logged in user")
-			}
-
-			authTokenResp, err := c.GetAuthToken(c.Ctx(), &auth.GetAuthTokenRequest{
-				Subject: whoamiResp.Username,
-			})
-			if err != nil {
-				return errors.Wrapf(grpcutil.ScrubGRPC(err), "could not get an auth token")
-			}
-
-			if jupyterhubChartVersion == "" {
-				jupyterhubChartVersion = getCompatibleVersion("jupyterhub", "/jupyterhub", defaultIDEChartVersion)
-			}
-			if hubImage == "" || userImage == "" {
-				ideVersion := getCompatibleVersion("ide", "/ide", defaultIDEVersion)
-				if hubImage == "" {
-					hubImage = fmt.Sprintf("%s:%s", defaultIDEHubImage, ideVersion)
-				}
-				if userImage == "" {
-					userImage = fmt.Sprintf("%s:%s", defaultIDEUserImage, ideVersion)
-				}
-			}
-
-			hubImageName, hubImageTag := docker.ParseRepositoryTag(hubImage)
-			userImageName, userImageTag := docker.ParseRepositoryTag(userImage)
-
-			values := map[string]interface{}{
-				"hub": map[string]interface{}{
-					"image": map[string]interface{}{
-						"name": hubImageName,
-						"tag":  hubImageTag,
-					},
-					"extraConfig": map[string]interface{}{
-						"templates": "c.JupyterHub.template_paths = ['/app/templates']",
-					},
-				},
-				"singleuser": map[string]interface{}{
-					"image": map[string]interface{}{
-						"name": userImageName,
-						"tag":  userImageTag,
-					},
-					"defaultUrl": "/lab",
-				},
-				"auth": map[string]interface{}{
-					"state": map[string]interface{}{
-						"enabled":   true,
-						"cryptoKey": generateSecureToken(16),
-					},
-					"type": "custom",
-					"custom": map[string]interface{}{
-						"className": "pachyderm_authenticator.PachydermAuthenticator",
-						"config": map[string]interface{}{
-							"pach_auth_token": authTokenResp.Token,
-						},
-					},
-					"admin": map[string]interface{}{
-						"users": []string{whoamiResp.Username},
-					},
-				},
-				"proxy": map[string]interface{}{
-					"secretToken": generateSecureToken(16),
-				},
-			}
-
-			if lbTLSHost != "" && lbTLSEmail != "" {
-				values["https"] = map[string]interface{}{
-					"hosts": []string{lbTLSHost},
-					"letsencrypt": map[string]interface{}{
-						"contactEmail": lbTLSEmail,
-					},
-				}
-			}
-
-			if dryRun {
-				var buf bytes.Buffer
-				enc := encoder(outputFormat, &buf)
-				if err = enc.Encode(values); err != nil {
-					return err
-				}
-				_, err = os.Stdout.Write(buf.Bytes())
-				return err
-			}
-
-			_, err = helm.Deploy(
-				activeContext,
-				"jupyterhub",
-				"https://jupyterhub.github.io/helm-chart/",
-				"pachyderm-ide",
-				"jupyterhub/jupyterhub",
-				jupyterhubChartVersion,
-				values,
-			)
-			if err != nil {
-				return errors.Wrapf(err, "failed to deploy Pachyderm IDE")
-			}
-
-			fmt.Println(ideNotes)
-			return nil
-		}),
-	}
-	deployIDE.Flags().StringVar(&lbTLSHost, "lb-tls-host", "", "Hostname for minting a Let's Encrypt TLS cert on the load balancer")
-	deployIDE.Flags().StringVar(&lbTLSEmail, "lb-tls-email", "", "Contact email for minting a Let's Encrypt TLS cert on the load balancer")
-	deployIDE.Flags().BoolVar(&dryRun, "dry-run", false, "Don't actually deploy, instead just print the Helm config.")
-	deployIDE.Flags().StringVarP(&outputFormat, "output", "o", "json", "Output format. One of: json|yaml")
-	deployIDE.Flags().StringVar(&jupyterhubChartVersion, "jupyterhub-chart-version", "", "Version of the underlying Zero to JupyterHub with Kubernetes helm chart to use. By default this value is automatically derived.")
-	deployIDE.Flags().StringVar(&hubImage, "hub-image", "", "Image for IDE hub. By default this value is automatically derived.")
-	deployIDE.Flags().StringVar(&userImage, "user-image", "", "Image for IDE user environments. By default this value is automatically derived.")
+	deployIDE := CreateDeployIDECommand()
 	commands = append(commands, cmdutil.CreateAlias(deployIDE, "deploy ide"))
 
 	deploy := &cobra.Command{
@@ -1171,231 +777,11 @@ func Cmds() []*cobra.Command {
 	}
 	commands = append(commands, cmdutil.CreateAlias(deploy, "deploy"))
 
-	var all bool
-	var includingMetadata bool
-	var includingIDE bool
-	var namespace string
-	undeploy := &cobra.Command{
-		Short: "Tear down a deployed Pachyderm cluster.",
-		Long:  "Tear down a deployed Pachyderm cluster.",
-		Run: cmdutil.RunFixedArgs(0, func(args []string) error {
-			// TODO(ys): remove the `--namespace` flag here eventually
-			if namespace != "" {
-				fmt.Printf("WARNING: The `--namespace` flag is deprecated and will be removed in a future version. Please set the namespace in the pachyderm context instead: pachctl config update context `pachctl config get active-context` --namespace '%s'\n", namespace)
-			}
-			// TODO(ys): remove the `--all` flag here eventually
-			if all {
-				fmt.Printf("WARNING: The `--all` flag is deprecated and will be removed in a future version. Please use `--metadata` instead.\n")
-				includingMetadata = true
-			}
-
-			if includingMetadata {
-				fmt.Printf(`
-You are going to delete persistent volumes where metadata is stored. If your
-persistent volumes were dynamically provisioned (i.e. if you used the
-"--dynamic-etcd-nodes" flag), the underlying volumes will be removed, making
-metadata such as repos, commits, pipelines, and jobs unrecoverable. If your
-persistent volume was manually provisioned (i.e. if you used the
-"--static-etcd-volume" flag), the underlying volume will not be removed.
-`)
-			}
-
-			if ok, err := cmdutil.InteractiveConfirm(); err != nil {
-				return err
-			} else if !ok {
-				return nil
-			}
-
-			cfg, err := config.Read(false)
-			if err != nil {
-				return err
-			}
-			_, activeContext, err := cfg.ActiveContext(true)
-			if err != nil {
-				return err
-			}
-
-			if namespace == "" {
-				namespace = activeContext.Namespace
-			}
-
-			assets := []string{
-				"service",
-				"replicationcontroller",
-				"deployment",
-				"serviceaccount",
-				"secret",
-				"statefulset",
-				"clusterrole",
-				"clusterrolebinding",
-			}
-			if includingMetadata {
-				assets = append(assets, []string{
-					"storageclass",
-					"persistentvolumeclaim",
-					"persistentvolume",
-				}...)
-			}
-			if err := kubectl(nil, activeContext, "delete", strings.Join(assets, ","), "-l", "suite=pachyderm", "--namespace", namespace); err != nil {
-				return err
-			}
-
-			if includingIDE {
-				// remove IDE
-				if err = helm.Destroy(activeContext, "pachyderm-ide", namespace); err != nil {
-					log.Errorf("failed to delete helm installation: %v", err)
-				}
-				ideAssets := []string{
-					"replicaset",
-					"deployment",
-					"service",
-					"pod",
-				}
-				if err = kubectl(nil, activeContext, "delete", strings.Join(ideAssets, ","), "-l", "app=jupyterhub", "--namespace", namespace); err != nil {
-					return err
-				}
-			}
-
-			// remove the context from the config
-			kubeConfig, err := config.RawKubeConfig()
-			if err != nil {
-				return err
-			}
-			kubeContext := kubeConfig.Contexts[kubeConfig.CurrentContext]
-			if kubeContext != nil {
-				cfg, err := config.Read(true)
-				if err != nil {
-					return err
-				}
-				ctx := &config.Context{
-					ClusterName: kubeContext.Cluster,
-					AuthInfo:    kubeContext.AuthInfo,
-					Namespace:   namespace,
-				}
-
-				// remove _all_ contexts associated with this
-				// deployment
-				configUpdated := false
-				for {
-					contextName, _ := findEquivalentContext(cfg, ctx)
-					if contextName == "" {
-						break
-					}
-					configUpdated = true
-					delete(cfg.V2.Contexts, contextName)
-					if contextName == cfg.V2.ActiveContext {
-						cfg.V2.ActiveContext = ""
-					}
-				}
-				if configUpdated {
-					if err = cfg.Write(); err != nil {
-						return err
-					}
-				}
-			}
-
-			return nil
-		}),
-	}
-	undeploy.Flags().BoolVarP(&all, "all", "a", false, "DEPRECATED: Use \"--metadata\" instead.")
-	undeploy.Flags().BoolVarP(&includingMetadata, "metadata", "", false, `
-Delete persistent volumes where metadata is stored. If your persistent volumes
-were dynamically provisioned (i.e. if you used the "--dynamic-etcd-nodes"
-flag), the underlying volumes will be removed, making metadata such as repos,
-commits, pipelines, and jobs unrecoverable. If your persistent volume was
-manually provisioned (i.e. if you used the "--static-etcd-volume" flag), the
-underlying volume will not be removed.`)
-	undeploy.Flags().BoolVarP(&includingIDE, "ide", "", false, "Delete the Pachyderm IDE deployment if it exists.")
-	undeploy.Flags().StringVar(&namespace, "namespace", "", "Kubernetes namespace to undeploy Pachyderm from.")
+	undeploy := CreateUndeployCmd()
 	commands = append(commands, cmdutil.CreateAlias(undeploy, "undeploy"))
 
-	var updateDashDryRun bool
-	var updateDashOutputFormat string
-	updateDash := &cobra.Command{
-		Short: "Update and redeploy the Pachyderm Dashboard at the latest compatible version.",
-		Long:  "Update and redeploy the Pachyderm Dashboard at the latest compatible version.",
-		Run: cmdutil.RunFixedArgs(0, func(args []string) error {
-			cfg, err := config.Read(false)
-			if err != nil {
-				return err
-			}
-			_, activeContext, err := cfg.ActiveContext(false)
-			if err != nil {
-				return err
-			}
-
-			// Undeploy the dash
-			if !updateDashDryRun {
-				if err := kubectl(nil, activeContext, "delete", "deploy", "-l", "suite=pachyderm,app=dash"); err != nil {
-					return err
-				}
-				if err := kubectl(nil, activeContext, "delete", "svc", "-l", "suite=pachyderm,app=dash"); err != nil {
-					return err
-				}
-			}
-
-			// Redeploy the dash
-			var buf bytes.Buffer
-			opts := &assets.AssetOpts{
-				DashOnly:  true,
-				DashImage: fmt.Sprintf("%s:%s", defaultDashImage, getCompatibleVersion("dash", "", defaultDashVersion)),
-			}
-			if err := assets.WriteDashboardAssets(
-				encoder(updateDashOutputFormat, &buf), opts,
-			); err != nil {
-				return err
-			}
-			return kubectlCreate(updateDashDryRun, buf.Bytes(), opts)
-		}),
-	}
-	updateDash.Flags().BoolVar(&updateDashDryRun, "dry-run", false, "Don't actually deploy Pachyderm Dash to Kubernetes, instead just print the manifest.")
-	updateDash.Flags().StringVarP(&updateDashOutputFormat, "output", "o", "json", "Output format. One of: json|yaml")
+	updateDash := CreateUpdateDashCmd()
 	commands = append(commands, cmdutil.CreateAlias(updateDash, "update-dash"))
 
 	return commands
-}
-
-// getCompatibleVersion gets the compatible version of another piece of
-// software, or falls back to a default
-func getCompatibleVersion(displayName, subpath, defaultValue string) string {
-	var relVersion string
-	// This is the branch where to look.
-	// When a new version needs to be pushed we can just update the
-	// compatibility file in pachyderm repo branch. A (re)deploy will pick it
-	// up. To make this work we have to point the URL to the branch (not tag)
-	// in the repo.
-	branch := version.BranchFromVersion(version.Version)
-	if version.IsCustomRelease(version.Version) {
-		relVersion = version.PrettyPrintVersionNoAdditional(version.Version)
-	} else {
-		relVersion = version.PrettyPrintVersion(version.Version)
-	}
-
-	url := fmt.Sprintf("https://raw.githubusercontent.com/pachyderm/pachyderm/compatibility%s/etc/%s/%s", branch, subpath, relVersion)
-	resp, err := http.Get(url)
-	if err != nil {
-		log.Warningf("error looking up compatible version of %s, falling back to %s: %v", displayName, defaultValue, err)
-		return defaultValue
-	}
-
-	// Error on non-200; for the requests we're making, 200 is the only OK
-	// state
-	if resp.StatusCode != 200 {
-		log.Warningf("error looking up compatible version of %s, falling back to %s: unexpected return code %d", displayName, defaultValue, resp.StatusCode)
-		return defaultValue
-	}
-
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		log.Warningf("error looking up compatible version of %s, falling back to %s: %v", displayName, defaultValue, err)
-		return defaultValue
-	}
-
-	allVersions := strings.Split(strings.TrimSpace(string(body)), "\n")
-	if len(allVersions) < 1 {
-		log.Warningf("no compatible version of %s found, falling back to %s", displayName, defaultValue)
-		return defaultValue
-	}
-	latestVersion := strings.TrimSpace(allVersions[len(allVersions)-1])
-	return latestVersion
 }
